+++ src/ReviewPage.jsx
@@ -12,6 +12,7 @@ import CenteredCardLayout from "./components/CenteredCardLayout.jsx";
 import ReviewSection from "./components/ReviewSection.jsx";
 import { BOURSO_TRANSFER_PAYEE_ID } from "./constants.js";
 import { formStatePropType } from "./propTypes.js";
+import { formatYYYYMMDDLocal } from "./utils/dateUtils";
 import { getAccountIdByName } from "./utils/ynabUtils";
 
 export default function ReviewPage({ formState, onBack, onSubmitted }) {
@@ -25,7 +26,7 @@ export default function ReviewPage({ formState, onBack, onSubmitted }) {
   function createBaseTransaction(accountId, amount) {
     return {
       account_id: accountId,
-      date: new Date().toISOString().slice(0, 10),
+      date: formatYYYYMMDDLocal(formState.date),
       amount: amount,
       payee_id: formState.payeeId || null,
       payee_name: !formState.payeeId ? formState.payee : undefined,
@@ -148,14 +149,13 @@ export default function ReviewPage({ formState, onBack, onSubmitted }) {
       return;
     }
     const amount = (-formState.amountMilliunits / 1000).toFixed(2);
-    const now = Date.now();
     const tx = {
       category:
         formState.settleUpCategory === "∅"
           ? undefined
           : formState.settleUpCategory,
       currencyCode: formState.settleUpCurrency || "EUR",
-      dateTime: now,
+      dateTime: formState.date.getTime(),
       items: [
         {
           amount: amount,